+++ .eslintignore
@@ -1,3 +1,4 @@
 # Temporary workaround for Gulp's ESLint
 dist/
 node_modules/
+webpack.config.js
