+++ .yarnrc.yml
@@ -1,3 +1,10 @@
 nodeLinker: node-modules
 
 yarnPath: .yarn/releases/yarn-3.1.1.cjs
+
+npmScopes:
+    zuplo:
+        npmAlwaysAuth: true
+        npmAuthToken: '${GITHUB_NPM_TOKEN}'
+        npmPublishRegistry: 'https://npm.pkg.github.com'
+        npmRegistryServer: 'https://npm.pkg.github.com'
