--- lib/index.browser.ts
@@ -1,3 +1,0 @@
-export { default as Client } from './client.fetch';
-
-export * from './index';
