+++ lib/index.node.ts
@@ -21,6 +21,6 @@ export class IdentityVerification {
     }
 }
 
-export { default as Client } from './client';
+export { default as Client } from './client.node';
 
 export * from './index';
