+++ lib/index.worker.ts
@@ -1,0 +1,3 @@
+export { default as Client } from './client.worker';
+
+export * from './index';
