+++ package.json
@@ -12,13 +12,13 @@
     "files": [
         "dist"
     ],
-    "main": "dist/index.node.js",
-    "worker": "./dist/index.browser.js",
-    "browser": "./dist/index.browser.js",
+    "main": "./dist/lib/index.node.js",
+    "worker": "./dist/lib/index.worker.js",
+    "browser": "./dist/lib/index.worker.js",
     "exports": {
-        "browser": "./dist/index.browser.js",
-        "worker": "./dist/index.browser.js",
-        "node": "./dist/index.node.js"
+        "browser": "./dist/lib/index.worker.js",
+        "worker": "./dist/lib/index.worker.js",
+        "node": "./dist/lib/index.node.js"
     },
     "keywords": [
         "intercom",
@@ -46,14 +46,20 @@
         "nyc": "^15.1.0",
         "prettier": "^2.5.1",
         "sinon": "^7.3.2",
+        "ts-loader": "^9.2.8",
         "ts-node": "^10.4.0",
         "tslib": "^2.3.1",
-        "typescript": "^4.5.3"
+        "typescript": "^4.5.3",
+        "webpack": "^5.70.0",
+        "webpack-cli": "^4.9.2",
+        "webpack-node-externals": "^3.0.0"
     },
     "scripts": {
+        "build": "webpack",
         "clean": "rm -r -f dist",
         "static": "eslint .",
         "compile_ts": "tsc",
+        "compile_ts_worker": "tsc --project tsconfig.worker.json",
         "move_compiled_to_dist": "mv dist/lib/* dist && rmdir dist/lib",
         "prepublish": "yarn clean && yarn static && yarn compile_ts && yarn move_compiled_to_dist",
         "test:unit": "mocha -r ts-node/register test/unit/*.test.ts --reporter=nyan",
